+++ src/Entity/Callback/JobTimeoutCallback.php
@@ -4,6 +4,9 @@ declare(strict_types=1);
 
 namespace App\Entity\Callback;
 
+use webignition\BasilWorker\PersistenceBundle\Entity\Callback\CallbackEntity;
+use webignition\BasilWorker\PersistenceBundle\Entity\Callback\CallbackInterface;
+
 class JobTimeoutCallback extends AbstractCallbackWrapper
 {
     private int $maximumDurationInSeconds;
