+++ src/Event/TestExecuteDocumentReceivedEvent.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace App\Event;
 
-use App\Entity\Callback\CallbackInterface;
 use App\Entity\Callback\ExecuteDocumentReceivedCallback;
 use App\Entity\Test;
+use webignition\BasilWorker\PersistenceBundle\Entity\Callback\CallbackInterface;
 use webignition\YamlDocument\Document;
 
 class TestExecuteDocumentReceivedEvent extends AbstractTestEvent implements CallbackEventInterface
