+++ src/MessageHandler/SendCallbackHandler.php
@@ -4,32 +4,32 @@ declare(strict_types=1);
 
 namespace App\MessageHandler;
 
-use App\Entity\Callback\CallbackInterface;
 use App\Message\SendCallback;
-use App\Repository\CallbackRepository;
 use App\Services\CallbackSender;
-use App\Services\CallbackStateMutator;
 use Symfony\Component\Messenger\Handler\MessageHandlerInterface;
+use webignition\BasilWorker\PersistenceBundle\Entity\Callback\CallbackInterface;
+use webignition\BasilWorker\PersistenceBundle\Services\CallbackStateMutator;
+use webignition\BasilWorker\PersistenceBundle\Services\CallbackStore;
 
 class SendCallbackHandler implements MessageHandlerInterface
 {
     private CallbackSender $sender;
-    private CallbackRepository $repository;
+    private CallbackStore $callbackStore;
     private CallbackStateMutator $stateMutator;
 
     public function __construct(
         CallbackSender $sender,
-        CallbackRepository $repository,
+        CallbackStore $callbackStore,
         CallbackStateMutator $stateMutator
     ) {
         $this->sender = $sender;
-        $this->repository = $repository;
+        $this->callbackStore = $callbackStore;
         $this->stateMutator = $stateMutator;
     }
 
     public function __invoke(SendCallback $message): void
     {
-        $callback = $this->repository->find($message->getCallbackId());
+        $callback = $this->callbackStore->get($message->getCallbackId());
 
         if ($callback instanceof CallbackInterface) {
             $this->stateMutator->setSending($callback);
