+++ src/Services/CallbackSender.php
@@ -4,10 +4,11 @@ declare(strict_types=1);
 
 namespace App\Services;
 
-use App\Entity\Callback\CallbackInterface;
 use App\HttpMessage\CallbackRequest;
 use Psr\Http\Client\ClientExceptionInterface;
 use Psr\Http\Client\ClientInterface as HttpClientInterface;
+use webignition\BasilWorker\PersistenceBundle\Entity\Callback\CallbackInterface;
+use webignition\BasilWorker\PersistenceBundle\Services\CallbackStateMutator;
 use webignition\BasilWorker\PersistenceBundle\Services\JobStore;
 
 class CallbackSender
