+++ tests/Model/Entity/Callback/TestCallbackEntity.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace App\Tests\Model\Entity\Callback;
 
 use App\Entity\Callback\AbstractCallbackWrapper;
-use App\Entity\Callback\CallbackEntity;
-use App\Entity\Callback\CallbackInterface;
+use webignition\BasilWorker\PersistenceBundle\Entity\Callback\CallbackEntity;
+use webignition\BasilWorker\PersistenceBundle\Entity\Callback\CallbackInterface;
 
 class TestCallbackEntity extends AbstractCallbackWrapper implements CallbackInterface
 {
