+++ tests/Services/InvokableFactory/CallbackSetup.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Tests\Services\InvokableFactory;
 
-use App\Entity\Callback\CallbackInterface;
+use webignition\BasilWorker\PersistenceBundle\Entity\Callback\CallbackInterface;
 
 class CallbackSetup
 {
