+++ tests/Unit/HttpMessage/CallbackRequestTest.php
@@ -4,11 +4,11 @@ declare(strict_types=1);
 
 namespace App\Tests\Unit\HttpMessage;
 
-use App\Entity\Callback\CallbackInterface;
 use App\HttpMessage\CallbackRequest;
 use App\Tests\Mock\Entity\MockJob;
 use PHPUnit\Framework\TestCase;
 use Psr\Http\Message\RequestInterface;
+use webignition\BasilWorker\PersistenceBundle\Entity\Callback\CallbackInterface;
 
 class CallbackRequestTest extends TestCase
 {
